Update tuya.ts
Today I discovered that this dimmer module (_TZE204_5cuocqty' TS0601) also has the function for choosing the "switch type". This way you can indicate whether a toggle or momentary switch is connected. this is comparable to module _TZ3210_3mpwqzuu. Should I submit a change request for this? otherwise this function cannot be used.

I looked up the Tuya Datapoint via the Tuya gateway and see that the SwitchType datapoint has "57". I find it difficult to merge this with the T0601_dimmer_5 as I don't really know how to enter the other variables on this line. could you help me with this?